+++ exams/2014-15/2015-06-29/d3.py
@@ -49,7 +49,7 @@ def check_child(child, parent):
     """
 
     if child is not None:
-        return child < parent
+        return child <= parent
     else:
         return True